update rating policy

+++ app/controllers/api/v1/ratings_controller.rb
@@ -3,10 +3,11 @@ class Api::V1::RatingsController < Api::V1::BaseApiController
 
   def create
     # delete old rating if exist
-    old_rating = Rating.find_by(rater_id: current_user.id,
-                                rated_id: strong_params[:rated_id])
-    old_rating.destroy unless old_rating.nil?
     rating = Rating.new(strong_params.merge({ rater_id: current_user.id}))
+    authorize rating
+
+    old_rating = Rating.find_by(rater_id: current_user.id,rated_id:
+      strong_params[:rated_id]).destroy
     if rating.save
       render json: {
         message: t('.create_rating'),
@@ -19,6 +20,7 @@ class Api::V1::RatingsController < Api::V1::BaseApiController
 
   def destroy
     rating = Rating.find(params[:id])
+    authorize rating
     rating.destroy
     render json: { message: t('.destroy_rating') }, status: 200
   end
@@ -26,7 +28,7 @@ class Api::V1::RatingsController < Api::V1::BaseApiController
   private
 
   def check_permission_rate
-    unless current_user.courses_as_students.pluck(:student_id).include?(current_user.id)
+    unless current_user.courses_as_students.where(teacher_id: strong_params[:rated_id])
       render json: { message: I18n.t('.dont_have_permission') }, status: 406
     end
   end
